Exercise 3 for lab 3 completed.
Finalized the driver, however unsure if it was wanted in MortgageDriver or not. Reading the assignment it requested after all info is completed, it displays all mortgages. So in this case, driver was deleted and ProcessMortgage was finalized. Also corrected some minor errors from previous upload, tried doing decimals for percentages, however was not user friendly, so I changed it so where 1 = 1%, making it easier for the user to input the interest rate. Also changed some interest rate names to match across the java classes. Some smaller text corrections, and removed unneeded stuff in Mortgage. Should be completed.