Exercise 1 completed for Lab 4.
Creates a window utilizing a bunch of grids organized and works as intended. JavaFX was producing errors as I put it into the project as a whole instead of a module, so I corrected that and its working now. It is still hardcoded and giving errors, but not a big deal. Will update shortly.