Fixed JavaFX errors.
Created a new library folder called lib and moved the JavaFX lib folder into there. This allows for the project to be portable and isn't hardcoded to my computer. TLDR just means can download it on another computer and the JavaFX setup should be just fine. Also putting it into lib allows me to utilize it in future labs if needed and doesn't bloat other modules that don't need it. Does include a run configuration to also run the program so it doesn't produce errors.